refactor: 테스트에서만 사용하는 상수 테스트로 이동
