+++ src/main/kotlin/racingcar/util/NumberGenerator.kt
@@ -2,9 +2,4 @@ package racingcar.util
 
 interface NumberGenerator {
     fun getNumber(): Int
-
-    companion object {
-        const val MIN_GENERATED_NUMBER = 0
-        const val MAX_GENERATED_NUMBER = 9
-    }
 }
