+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -9,8 +9,8 @@ import io.kotest.matchers.shouldBe
 import racingcar.domain.Car.Companion.MOVE_THRESHOLD
 import racingcar.domain.Car.Companion.START_POSITION
 import racingcar.domain.Car.Companion.STEP
-import racingcar.util.NumberGenerator.Companion.MAX_GENERATED_NUMBER
-import racingcar.util.NumberGenerator.Companion.MIN_GENERATED_NUMBER
+import racingcar.util.Constant.MAX_GENERATED_NUMBER
+import racingcar.util.Constant.MIN_GENERATED_NUMBER
 import racingcar.util.OrderNumberGenerator
 
 class CarTest : BehaviorSpec({
