+++ src/test/kotlin/racingcar/racing/RacingTest.kt
@@ -6,8 +6,8 @@ import racingcar.domain.Car.Companion.MOVE_THRESHOLD
 import racingcar.domain.Car.Companion.START_POSITION
 import racingcar.domain.Car.Companion.STEP
 import racingcar.domain.RacingGameParam
-import racingcar.util.NumberGenerator.Companion.MAX_GENERATED_NUMBER
-import racingcar.util.NumberGenerator.Companion.MIN_GENERATED_NUMBER
+import racingcar.util.Constant.MAX_GENERATED_NUMBER
+import racingcar.util.Constant.MIN_GENERATED_NUMBER
 import racingcar.util.OrderNumberGenerator
 
 class RacingTest : BehaviorSpec({
