+++ src/test/kotlin/racingcar/util/Constant.kt
@@ -1,0 +1,6 @@
+package racingcar.util
+
+object Constant {
+    const val MIN_GENERATED_NUMBER = 0
+    const val MAX_GENERATED_NUMBER = 9
+}
